Add Authorization for both author and admin roles in post section and also in comment section. Apply bug fixes pertaining to update post and update comment feature. Added Draft page. Applied AUthorization to Draft page as well.

+++ src/main/java/com/shantanu/blogapp/controller/CommentController.java
@@ -1,10 +1,15 @@
 package com.shantanu.blogapp.controller;
 
+import com.shantanu.blogapp.config.UserDetailsImpl;
 import com.shantanu.blogapp.entity.Comment;
 import com.shantanu.blogapp.entity.Post;
+import com.shantanu.blogapp.entity.User;
+import com.shantanu.blogapp.repository.UserRepository;
 import com.shantanu.blogapp.service.CommentService;
 import com.shantanu.blogapp.service.PostService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -21,21 +26,31 @@ public class CommentController {
 	@Autowired
 	private PostService postService;
 
+	@Autowired
+	private UserRepository userRepository;
+
 	@PostMapping("/post/saveComment/{id}")
-	public String saveComment(@PathVariable("id") int id, Comment comment) {
+	public String saveComment(@PathVariable("id") int id, Comment comment,
+														@AuthenticationPrincipal UserDetailsImpl currentUser) {
 		Post post = postService.getPostById(id);
-		post.addComment(commentService.addCommentDetails(post, comment));
+		post.addComment(commentService.addCommentDetails(post, comment, currentUser));
 		postService.saveComment(post);
 		return "redirect:/post/{id}";
 	}
 
 
 	@GetMapping("/post/{postId}/comment/{commentId}")
-	public String showupdateCommentForm(@PathVariable("commentId") int commentId,
+	public String showUpdateCommentForm(@PathVariable("commentId") int commentId,
 																			@PathVariable("postId") int postId,
 																			@ModelAttribute("post") Post post,
 																			@ModelAttribute("comment") Comment comment,
+																			@AuthenticationPrincipal UserDetails currentUser,
 																			Model model) {
+		Post oldPost = postService.getPostById(postId);
+		User user = userRepository.findByUsername(currentUser.getUsername()).get();
+		if(!currentUser.getUsername().equals(oldPost.getAuthor()) &&  !user.getRole().equals("ROLE_ADMIN")) {
+			throw new RuntimeException("You are not authorized to view this page");
+		}
 		post.setId(postId);
 		Comment oldComment = commentService.getCommentById(commentId);
 		commentService.getOldComment(comment, oldComment);
@@ -51,7 +66,14 @@ public class CommentController {
 	}
 
 	@PostMapping("/post/{postId}/deleteComment/{commentId}")
-	public String deleteComment(@PathVariable("commentId") int commentId) {
+	public String deleteComment(@PathVariable("postId") int postId,
+															@PathVariable("commentId") int commentId,
+															@AuthenticationPrincipal UserDetails currentUser) {
+		Post post = postService.getPostById(postId);
+		User user = userRepository.findByUsername(currentUser.getUsername()).get();
+		if(!currentUser.getUsername().equals(post.getAuthor()) &&  !user.getRole().equals("ROLE_ADMIN")) {
+			throw new RuntimeException("You are not authorized to view this page");
+		}
 		Comment comment = commentService.getCommentById(commentId);
 		commentService.deleteComment(comment);
 		return "redirect:/post/{postId}";
